fix(engine): Fix object store client initialization for data object
The client for reading data objects from object store uses the new
Thanos client library. When using a named store, this initialization
fails, because it does not map the store name to the store backend.

This commit also adds the thanos named store config to the integration
test setup, because named stores for Thanos need to be configured
separately from the legacy `named_store` block.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>