chore: refactor how distributor checks ingest-limits
This commit refactors how the distributor checks ingest-limits to
prepare for performing different actions depending on which limits
are exceeded. For example, if a stream limit is exceeded then we
shouldn't fail the request. On the other hand, if the per-tenant
data rate limit is exceeded, the entire request should be failed.
This code will be added in a subsequent commit.