chore: align benchmark results between chunks and dataobjs
There were some small inconsistencies between chunks and dataobjs when
evaluating the benchmark queries, making it hard to evaluate whether the
becnhmark truly reflected performance differences between the two
storage formats.

This commit adds a test to check the output of each query for each
store, and compare it against the expected baseline (chunks).

Because this test is slow and requires generating the data, a new
-slow-tests flag has been added to the pkg/logql/bench package. This
test will be ignored if that flag isn't present.

This also means that CI will not run these tests by default, and we
would need to set up a new pipeline if we wanted them to be ran (which
would include generating the test data).