feat: Optionally enable kafka histograms for read/write latency
Introduces a new config option for kafka clients to enable (or disable)
read/write wait and timing histograms.  These are not enabled by
default, so this change introduces a new CLI flag and YAML opt which
configures the collected metrics to include/exclude historgams.

This PR includes updating the dataobj, blockbuilder (deprecated),
and ingest limits so they can take advantage of these collected metrics
for monitoring.