chore(logql/bench): store proper timestamps in generated chunks
The timestamp bounds of a chunk are used to update the query range
bounds; the previous code always truncated the bounds down, which caused
backwards queries to ignore the latest log line in a chunk if that log
line was less than a millisecond beyond the truncated timestamp.

PR #17104 was inadvertently running generated code with this fix already
present, but the fix was removed after I thought it wasn't important.

Turns out it was important :)