fix: fix a bug where limits were incorrect after grafana/loki#16937
This commit fixes a bug where limits were incorrect after the change
in grafana/loki#16937. This bug occurred because we stopped asking
each limit instance if it knew about all stream hashes, and instead
starting asking each limit instance if it knew about the stream
hashes for its assigned partitions. This broke the logic in the
frontend that was taking the intersection of all responses to
calculate if a stream was unknown. It should now take the union
instead.