fix(engine): Fix querier intitalisation
The recent wiring of the v2 query engine in the querier caused a
regression in the initialisation of the querier when the Thanos object
client was not configured, because the v2 engine uses DataObject and
DataObjects rely on Thanos object client.

This commit fixes the problem by creating the Thanos object client only
when v2 engine is enabled.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>