chore(engine): Implement execution pipeline for SortMerge operator
KWayMerge is a k-way merge of multiple sorted inputs.
It requires the input batches to be sorted in the same order (ASC/DESC) as the SortMerge operator itself.
The sort order is defined by the direction of the query, which is either FORWARD or BACKWARDS,
which is applied to the SortMerge as well as to the DataObjScan during query planning.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>