feat: refactors kafka/client to improve metric collection
Updates the `pkg/kafka/*.go` code base to refine how we're initializing
Kafka clients and configuring metrics.  This set of changes removes
functionality for calling services to inject `kprom.Metrics` given all
services are relying on the base set.

In doing so, this set of changes improves the metric names so they
represent the following:
* The metrics are collected by the Kafka client (`loki_kafka_client_`)
* The metrics are collected by a reader/writer (label: "service_role: reader|writer")
* The metrics are collected by a particular component (label:
  "component: ...")

This set of changes also adds a few metrics to improve our coverage, and
it alters one metric description to help better explain what we're
collecting.

In the future, we can easily extend this to allow services to inject
customized `kprom.Metrics` structs if needed.