feat: support disabling the partition consumers cache
This commit updates caching in the limits-frontend to support
disabling the partition consumers cache via a no-op cache. We use
a generic interface Cache[K comparable, V any] and provide two
implementations, a TTLCache and a NopCache. The TTLCache is a
much simpler implementation from our previous cache, and does not
require goroutines to manage cache eviction. It does not support
LRU as we do not need it.