chore: adjust metric names that were missed
The distributor code was missing additional identification for the
service and its role.  This commit adds an additional function to the
kafka writer to wrap a prometheus registerer with the appropriate
labels.  These labels are applied to the code the distributor uses to
create a new producer client.

Also added two constants for the `service_role` label.