feat: update loki canary to support batched log push
This commit introduces a new `EntryWriter` implementation in the Loki
canary codebase, supporting batched log pushes.  A new runtime CLI opt,
`-log-batch-size` receives an int, specifying the # of log lines to
batch before they are sent to Loki.

This opt may be set to '0' or '1' to disable batching entirely, in which
case the canary functionality remains as-is.

If an int value > 1 is passed to this CLI opt, a new `BatchedPush`
instance is created.  This instance collects log lines from its channel
and stores them in local array; once the array size is at least the
value of `-log-batch-size`, the logs are then marshaled and pushed to
Loki using the standard push API.

This implementation also maintains a 30s countdown timer; every 30s, the
timer triggers and whatever logs exist in the "batch" array are pushed
to Loki.  This prevents a slowdown in log ingestion into the canary
instance from preventing logs from backing up for too long.

Moreover, if/when the canary org is shutdown (i.e., the termination
channel receives a signal to quit), the remaining logs are likewise
pushed to Loki and the shutdown process continues.