chore: remove the lockable array for a simpler lock mechanism
Rather than have a lockable struct with the log lines array, just
maintain a simple callback mechanism which implements locking.  This
means every time a channel in the `BatchedPush.run` function receives a
response, run a function which acquires the lock, checks the array size
(if applicable), flushes all batched log lines to Loki, and releases the
lock.

Either of these implementations should work, but this is simpler.