chore: restructure some of our limits metrics
This commit restructures some of our limits metrics. It removes
tenant from _ingest_limits_partitions as this was redundant
information. The same pod cannot consume different partitions
for different tenants. It also counts the number of active and
expired streams per tenant, and removes the recorded metric
which was not being used.