feat: add kafkav2 client
This commit adds a new kafka client called clientv2. I will replace
existing use of the current package over a number of subsequent
PRs to make it easier to review and test changes in isolation.

The motivation for this new package is that it was hard to create
a *kgo.Client that could both produce and consume records. To fix
this, the v2 client separates creating the client from creating
the options, enabling the use of clients that produce, consume,
and do both.