refactor(dataobj): bound section decoders to a section
Originally, instances of `encoding.StreamsDecoder` and
`encoding.LogsDecoder` could be used to read columns and pages from any
section of those types.

However, now that sections can specify their data regions, and offsets
are relative to the start of the data region (when specified), we have a
need to scope instances of StreamsDecoder and LogsDecoder down to a
single section.

This commit updates StreamsDecoder and LogsDecoder to be given a section
upon creation. A future commit will make use of data regions.