chore(dataobj): support reading section layout
This commit adds support to using the section layout:

* When the layout is specified, metadata is read from the layout. An
  error is returned if the layout is specified alongside the deprecated
  fields.

* When the layout is specified, reading data fails if a data region
  isn't specified. Reads of section data is relative to the start of the
  region.

This change is a no-op: nothing is writing the layout, so the old
behaviour is used. Making this change a no-op allows us to test this
commit in isolation to verify that the fallback behaviour is correct.
This will be done in lieu of new tests, which would otherwise require
support for writing both the new and old format.