chore(dataobj): write objects with section layout information
This commit migrates encoding to encode section layout information:

* The offset and length of the data region is written

* The deprecated fields (MetadataOffset, MetadataSize) are no longer
  written.

* Streams and logs sections now use section-relative offsets.

As a result of this change, our test objects are different. This causes
the metastore to write them in a slightly different order, breaking
tests. I've updated these tests to attempt to remove the constraint on
how objects are written to decouple them.

With this change:

* All new objects will be written using SectionLayout for indicating
  where the data region of a section is.

* Old objects can still be read.