chore(dataobj): rename SectionType to SectionKind
Whenever a new section is created, the filemd protobuf must be updated
to note the new kind of section. Updating the top-level file metadata
to be aware of all sections is a small (but surprising) burden that
developers have to undertake. Additionally, it prevents dataobjs from
ever being used as a library without having a global registry of all
known sections.

Hard-coded section types will be replaced with a file-scoped registry of
named types. To make way for this change, the previous SectionType is
renamed to SectionKind.