refactor(dataobj): Update encoding.Decoder to expose Metadata
Previously, encoding.Decoder exposed a subset of the filemd.Metadata
type (specifically the sections). Now that filemd.Metadata includes
more than just a list of sections, Decoders should expose all of the
metadata for callers to use.

The alternative approach would be to add new methods for each subset of
Metadata to retrieve, but each method call would be a round trip request
and re-parsing the same metadata, so it's more efficient to have the
more general method.