chore(dataobj): add shared pool for unbucketed buffer pool
bufpool was originally intended to be used when the size of the buffer
is known, but this doesn't happen for encoding. As we're going to move
section encoders to their own packages, we first need to do some prep
work so that packages can share access to the pool of unsized
bytes.Buffer.