refactor(dataobj): move StreamEncoder to streams package
This moves the StreamsEncoder to the streams package. The test for the
StreamsEncoder is no longer needed, as the overall test in the streams
package handles testing that code path.

As a side effect of this change, some types and variables needed to be
exported. Some of these exports will also move to the section packages
once all encoding and decoding functionality is split out.