refactor(dataobj): move LogsEncoder to logs package
This moves the LogsEncoder to the logs package. The test for the
LogsEncoder is no longer needed, as the overall test in the logs package
handles testing that code path.

As a side effect of this change, some types and variables needed to be
exported. Some of these exports will also move to section packages once
all encoding and decoding functionality is split out.