refactor(dataobj): keep Encoder for the lifetime of a Builder
Retaining an Encoder instance for the lieftime of a Builder allows us to
encode sections while a Builder is still accumulating data rather than
only encoding at flush time.

This is part of a change to have callers be responsible for when a
section is cut, and to constrain each section builder to only building a
single section.