feat: add CLI arg for canary batched pusher max limit
Adds a CLI arg w/a default value to limit the upper bound of the
loki-canary log pusher batch size.  Canary pods don't have a lot of
memory, so this implements a limit of 20 log lines in a given batch,
the goal being to restrict the memory to 20MB of logs in memory at any
given point in time.

This limit is only imposed in the canary's `main.go` module when
initializing the canary.  We don't do any additional checks in the
`batched_push.go` implementation to restrict the batch size to an
upper-bound.

This also reorganizes the push tests a little.