test: adds timeout canary batched push test w/minor changes
Fixes the existing terminate test so it accurately tests shutting down
the BufferedPush writer.  This also adds a separate test for ensuring
the timeout countdown effectively sends all logs in the batched array.

Updates the `main.go` module which checks the logBatchSize and Max so we
allow the batch size to be equal to the max size.

Also adds a few very minor changes to the `batched_push.go` module: a
few comment changes, moving some things around, removing an unnecessary
log line, and updating the `Stop()` function.