chore: add back deprecated config to avoid possible rollout issue
This commit adds back the window_size and bucket_duration fields as
no-ops. This is to avoid a possible problem we might have next week
with the releases where the ConfigMap is updated, and some pods
on the previous release start crash looping before they can be
rolled out to the new image.