ci: add workflow to manually run LogQL benchmarks
Add a new workflow to run the LogQL benchmarks in `./pkg/logql/bench`.
The workflow must be manually triggered due to how long the benchmarks
take to complete.

The workflow can be customized to specify which subset of stores to
benchmark; by default, both the chunk and dataobj v1 stores will be
benchmarked.

Since benchmarks tests 2GiB of generated testdata, the benchmarks run on
a [large runner].

Stores are benchmarked in parallel to reduce the amount of time waiting
for results. Each benchmark is ran 10 times for future comparison with
benchstat.

At the end of the benchmark runs, each store uploads an artifact
containing the benchmark results (as text), along with CPU and memory
profiles.

[large runner]: https://docs.github.com/en/actions/using-github-hosted-runners/using-larger-runners/about-larger-runners