feat: quick fix to produce a stream once per minute
We decided that producing a record on each update was too much,
and created a lot of records to replay on start up on partition
rebalance. Instead, what if we produce a record for a stream
if it has been 1 minute or longer since it was last produced.
This commit makes the absolute minimum changes required to validate
this. It accepts that rate buckets will be incorrect as not all
sizes are produced, which will be solved in the next phase of the
limits service.