chore(engine): add record-based topk implementation
This adds an implementation of topk which works by retaining records
that have at least one row which contributes to the topk.

If the total number of rows in retained records that do not contribute
ot the topk exceed a configurable amount, all retained records are
compacted into a smaller record that only contains the current heap.

The implementation is split into several components to make the code
easier to follow:

* arrowagg.Arrays allows aggregating arrow.Arrays together.
* arrowagg.Records allows aggreagting arrow.Records together.
* arrowagg.Mapper allows quickly locating fields in incoming schemas.

The final record has a combined schema, composed of the union of all
incoming schemas. topk inserts null values in rows where a field did not
exist from the original record.

This does not yet include unit tests for the various pieces of
functionality.