feat: add ReasonFailed
This commit adds a new reason called ReasonFailed. It is used in
proto.ExceedsLimitsResult when a stream cannot be checked against
a tenant's limits due to an error. It allows this information to
be fed back to the distributors which can then make a decision
whether to accept or reject the streams (for example, having a
per-tenant toggle).