chore(engine): Filter by unknown column (#18131)
The new engine fails to execute a query when an `arrow.Record` (batch) does not contain the column on which a label filter is used, e.g. `| foo="bar"` where `foo` is neither a label nor a structured metadata.

This change also fixed an incorrect conversion of the label filter operator `=`, which was converted to `MATCH_STR`, but requires to be `EQ`.

The new engine reflects current behaviour of a label filter `| foo=""` where `foo` is undefined.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>