fix: race condition where max attempts is never satisified
This commit fixes a race condition where the maximum attempts is
never satisifed. This race condition comes about as two (or more)
goroutines can load and increment the atomic concurrently, which
would cause the next Load() to load a value greater than the
maximum number of attempts, breaking the equality check.