fix: fix race condition with sync.Mutex
This commit fixes the race condition that exists between two
(or more) goroutines concurrently executing CheckReady. What can
happen is both goroutines call Load at the same time, do the check,
and then increment the number of attempts without using check-and-set.
This allowed the number of attempts to exceed the maximum number
of attempts.