fix(pattern): Fix feature flag for enabling pattern persistence (#18215)
PR https://github.com/grafana/loki/pull/17737 introduced a feature to persist detected patterns on the pattern ingesters. And while there is a dedicated global config `PatternPersistence` and a per-tenant config `PatternPersistenceEnabled`, none of these were honoured, because for writing the patterns to the distributor, the same writer (and its config) as for persisting aggregated metrics is used. So, when aggregated metrics is enabled, also patterns were automatically written/persisted as well.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>