feat: Mark truncated log lines with identifier (#18262)
While truncated JSON log lines are usually easy to identify (because the JSON becomes invalid), it is not always the case with other plain text log lines.

This PR adds the possibility to add an identifier as suffix for truncated log lines, such as `[...]` so that truncated log lines are easier to spot and identify when querying them.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>