chore(engine): Scan dataobj rows in batches
This commit changes the behaviour of the dataobj scan to read and yield
only a single batch of rows when the `Read()` function of the pipeline
is invoked. This is now possible because rows are sorted by timestamp in
descending order and the planner creates a single dataobjscan node per
section, which feed into a k-way merge (sortmerge) operator.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>