chore(engine): Fix logical plan for metric queries (#18471)
The logical plan for metric queries previously omitted the SORT instruction, because if would have translated into a SortMerge with a topK.

However, due to the recent change, the sort merge does not do the topK any more, but is still used to merge multiple DataObjScan inputs in DESC timestamp order.
This is required for the SELECT instruction (Filter operation), since it only accept a single input, which is ensured by the SORT.

Fixes the error `filter expects exactly one input, got N`.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>