chore(util/arrowtest): make arrowtest more ergonomic with timestamps
Previously, arrowtest could not be used to decode timestamps encoded
with arrowtest.Rows.Record:

* No Go type mapped to an Arrow timestamp datatype
* Timestamps always decoded back into a string representation

While this wasn't an issue when using detected schemas (nothing
converted into a timestamp), it becomes a problem when known schemas are
used.

This commit changes arrowtest to allow use of time.Time to convert to
and from Arrow timestamps:

* Detected schemas convert time.Time into the equivalent Arrow DataType
  at nanosecond precision (retaining time zone).

* An array of timestamps back into Go values now convert back into
  time.Time (retaining time zone of the array).

Now timestamps can be used much more naturally, and the timestamp string
helper is no longer needed.