chore(engine): add topk Pipeline
The topk Pipeline wraps around the record-based topkBatch
implementation, providing a SORT+LIMIT implementation from a set of
input pipelines.

This Pipeline is not intended to be a direct part of the query graph
(though it could be); rather, its primary use is as a mini Pipeline to
incorporate into larger Pipelines that are a part of the graph (such as
DataObjScan).

Using the pipeline this way allows for nodes like DataObjScan to easily
provide sorting when sorting is needed for a section.