chore(logproto): use new labels.Labels API
Use the new labels.Labels API when interacting with
logproto.LabelAdapter.

This change is likely to have some small regression on CPU and
allocations: it is no longer possible to use unsafe to cast a
[]logproto.LabelAdapter to labels.Labels.

The new implementation of the conversion functions works similarly to
the equivalent conversion functions in Prometheus, where memory is
reused wherever possible.

The resulting functions are still more expensive than before, but the
overhead should be controlled.

If memory allocations from using FromLabelsToLabelAdapters becomes too
high, the new CopyToLabelAdapters functions can help pay down this cost.