chore(engine): Add tests for predicate pushdown for ambiguous columns
Ambiguous columns (columns that can either be a label or structured
metadata) cannot be pushed down to the dataobjscan, because the dataobj
reader lacks the capability of handling the precedence of metadata over
labels.

This change adds test cases that ensure that predicates with ambiguous
columns are not pushed down in the optimizer.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>