chore(engine): add columnar implementation of DataObjScan
To simplify the new implementation, columns are no longer returned in
the order matching projections (when provided). Instead, columns are
ordered by:

1. Stream labels
2. Log metadata
3. Log timestamp
4. Log message

Additonally, DataObjScan will now force a topk sort if there is no known
sort information in the logs section, or if the sort order doesn't match
what's expected.