chore(dataobj): make copy of pages from downloaded window
Previously, pages were subslices of the entire download window. This
meant that as long as one of those pages was still reachable, the entire
window (up to 16MB) would remain active in memory.

Making copies out of the window will allow pages to be reclaimed by the
GC faster when under pressure, and avoid holding onto memory for
unneeded regions of the window.

To counteract the increased allocations when calling ReadPages, we now
download window into a pooled slice (but still allocate the byte slices
for individual pages).