chore(helm): Refactor existing loki health probes
This change refactors all existing uses of `.Values.loki.readinessProbe`
and `.Values.loki.livenessProbe` so that they treat the value `{}` as
meaning that the probe should be disabled.

Some components used `.Values.loki.livenessProbe` despite it not being
set in values.yaml. So if that value gets introduced with `{}` as the
default value, then it would break things without this refactoring.

Signed-off-by: Andreas Lindhé <7773090+lindhe@users.noreply.github.com>