chore(engine): don't sort metric queries
Bypassing SortMerge for metric queries appears to significantly speed up
query time.

This commit changes the default sort order from ASC to the new value
UNSORTED, and removes applying sort for metric queries.

To support this change, the physical planner needed to be updated so
that a SortMerge doesn't always wrap DataObjScan nodes.