chore(dataobj): return Object on Builder.Flush
This commit updates Builder.Flush to return an Object rather than
writing bytes to an argument. To support this change, dataobj.Object has
two additional methods:

- dataobj.Object.Size, to retrieve the full encoded size of the object
- dataobj.Object.Reader, to read from the encoded object

This API change allows downstream users of dataobj.Builder to stream the
encoded object for analysis or uploading, without requiring the caller
to retain the entire object in memory at once.

In the future, this will be used to reduce the memory overhead of
dataobj-consumers by buffering flushed sections on an ephemeral disk
instead of in memory.