test(kafka/partitionring/consumer): fix test flakiness
TestPartitionMonitorRebalancing was suceptible to "soft" deadlocks where
both committing offsets and processing revoked partitions were waiting
for the other to finish (via the WaitGroup used in the test).

In this case, "soft" deadlock refers to how some goroutines would be
blocked forever, but franz-go library managed to work its way around it
(by spawning other goroutines, aborting a wait on a callback, etc.)

In at least one case on CI, this deadlock appeared to be harder, where
the test timed out instead.

This commit attempts to fix the problem by instead doubling down on
commiting uncomitted offsets at the time partitions are revoked.

I've been running this fix in a loop for the past hour, and it no longer
produces a situation where the test takes longer than 30 seconds to
finish (on my machine), where it previously could randomly take 1-5
minutes.