feat(helm): Allow enabling user namespaces
This gates the ``.spec.hostUsers` field behind a semver version check which verifies that a Kubernetes cluster is at least on version 1.33. If that is the case, it sets the field to `true` which disables user namespaces by default because cluster administrators need to prepare their cluster nodes first, before they are able to switch on user namespaces.

Fixes #18646

Signed-off-by: Sebastian Hoß <seb@xn--ho-hia.de>