chore(dataobj): add sectionScratchStore abstraction
sectionScratchStore is an interface for representing how completed
sections are buffered while the rest of the data object is still being
created.

scratchStore has two implementations:

- memoryScratchStore, which accumulates completed sections in memory
- diskScratchStore, which accumulates completed sections on disk

diskScratchStore is intended to be given an ephemeral storage location.
Any existing files in the scratch filepath are ignored.

To handle situations where a disk runs out of space, diskScratchStore
will fallback to using in-memory storage if a write happens to fail.