chore(deps): Downgrade kprom from v1.3.0 back to v1.2.1
With https://github.com/grafana/loki/pull/18596, we upgraded kprom to a version that contains a breaking change.

The library now introduces a metric `{ns}_buffered_produce_bytes_total` which is already defined as part of both Mimir and adaptime metrics.

Rolling forward in Mimir is more complicated than rolling backwards here in Loki as I can see there are no dashboard or alerts using this metric so I propose to do that.

I'll also need to tell renovate to skip this version afterward

Signed-off-by: gotjosh <josue.abreu@gmail.com>