chore(loki): add scratch.Store abstraction
scratch.Store is a new Loki module which allows other modules to share a
common implementation for storing temporary scratch data.

scratch.Store has two implmeentations:

- scratch.Memory, which accumulates temporary data in memory
- scratch.Filesystem, which accumulates temporary data on the
  filesystem.

scratch.Filesystem is intended to be given an ephemeral storage
location. scratch.Filesystem writes files with random filenames and with
a ".lokiscratch" extension. To avoid disk memory usage leaking as a
result of Loki unexpectedly terminating, all files in the scratch
directory with the .lokiscratch extension are removed on module startup.

To handle situations where a disk runs out of space, scratch.Filesystem
falls back to an in-memory store if a write happens to fail.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>