chore(dataobj): permit reading subset of section metadata
Originally, sections were intended to read their entire metadata region
in a single call. This had the extra requirement for metadata to be as
small as possible and to only include what's always necessary.

As a result of this, sections based on dataset stored additional
metadata (like lists of pages per column) in the data region, allowing
the sections to retrieve that optional metadata on demand.

This decision caused some issues:

- Metadata is now fragmented across a section, making it difficult to
  minimize the overhead of downloading metadata from object storage

- The metadata region size histogram is not representative of the true
  metadata size, as some metadata is stored in the data region.

This commit changes the dataobj.SectionReader interface to permit a
section to read a subset of metadata. This will allow new format
versions of dataset to be introduced that store optional metadata in the
same metadata region.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>