chore(dataobj): remove support for deprecated fields
Back in May, PR #17704 introduced section-relative offsets, while PR
fields in the original protobuf for the older forms, and backwards
compatible code was introduced to continue reading these older objects.

Since it's been three months, it's a good opportunity to fully remove
these deprecated fields and the backwards compatibility.

With this change, attempting to read older objects will report one of
the following errors, depending on when it was made:

* `section has no data` and `section has no metadata`, for objects
  created before PR #17704.
* `missing section type: typeRef 0 out of bounds`, for objects created
  before PR #17708.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>