chore(dataobj): allow sections to specify custom info at the file level
With the introduction of PR #18815, sections may now partially read
their metadata. This allows sections to store much more information in
the metadata region without cost of opening a section.

However, if a section were to take advantage of this, it would have no
way of knowing what the smallest readable unit in the metadata region
is. As a workaround, it could attempt to read some "safe" fixed-size
portion of the metadata region, but:

* If that portion isn't large enough, a second round-trip to object
  storage is required, or
* the portion is too large, leading to wasted downloads.

This commit offers a utility for sections to avoid this issue by
encoding "extension data" to the file-level SectionInfo.

For example, a section may wish to encode a serialized protobuf message
like the following:

    message SectionExtensionInfo {
      // Expose the smallest readable unit of metadata

      uint32 metadata_start_offset = 1;
      uint32 metadata_start_length = 2;
    }

Then, since this info is provided at the file-level, sections can
immediately use this information and only read the exact amount of
metadata required when opening the section.

Since SectionInfo extensions are stored at the file-level metadata, this
functionality must be used sparingly, typically only for information
that must be known before reading metadata without an additional
roundtrip to storage.